signupform function moved to helper function
