+++ src/Components/SignupForm/index.js
@@ -5,358 +5,9 @@ import React from 'react';
 import './Signupform.css';
 import LoginSideAfterRegister from '../LoginSideAfterRegister';
 import QReader from '../QReader';
-
+import helpers from '../../helpers/signUpFormHelpers'
 const strftime = require('strftime');
 
-const getAadhaarForm = state => (
-  <div>
-    <input
-      type="text"
-      placeholder="12-digit aadhaar number"
-      className="Login-input-field1"
-      value={state.state.aadhaarNo}
-      onChange={(event) => {
-        if (event.target.value.match(/^[0-9]{0,12}$/)) {
-          state.setState({
-          aadhaarNo: event.target.value,
-          aadhaarError: '',
-        });
-      }
-      }}
-    />
-    <div className="Error-message">{state.state.aadhaarError}</div>
-  </div>
-);
-
-const verifyOTPForm = state => (
-  <div>
-    <input
-      type="password"
-      placeholder="_ _ _ _"
-      value={state.state.otp}
-      className="Login-input-field1"
-      onChange={(event) => {
-      if (event.target.value.match(/^[0-9]{0,4}$/)) {
-        state.setState({
-        otp: event.target.value,
-        otpError: '',
-      });
-    }
-    }}
-    />
-    <div className="Error-message">{state.state.otpError}</div>
-  </div>
-);
-
-const userRegisterForm = state => (
-  <div>
-    <input
-      type="text"
-      placeholder="Username"
-      className="Login-input-field1"
-      value={state.state.username}
-      onChange={(event) => {
-      if (event.target.value.match(/^[-\w.$@*!]{0,15}$$/)) {
-        state.setState({
-          username: event.target.value,
-          usernameError: '',
-        });
-      }
-    }}
-    />
-    <div className="Error-message">{state.state.usernameError}</div>
-    <input
-      type="password"
-      placeholder="Password"
-      className="Login-input-field1 Signup-field"
-      value={state.state.password}
-      onChange={(event) => {
-      state.setState({
-        password: event.target.value,
-        passwordError: '',
-      });
-  }}
-    />
-    <div className="Error-message">{state.state.passwordError}</div>
-  </div>
-);
-
-const progressBar = () => (
-  <div className="Signupform-progressbar">
-    <div className="Signupform-progressbar-dot" />
-    <div className="Signupform-progressbar-bar" />
-  </div>
-);
-
-const progressBarHalf = () => (
-  <div className="Signupform-progressbar">
-    <div className="Signupform-progressbar-dot" />
-    <div className="Signupform-progressbar-bar-unfilled" />
-  </div>
-);
-
-const sendOTP = state => (
-  <div className="Signupform-welcome-message">
-    <div className="Signupform-heading">Give Aadhaar Number</div>
-    {/* <div className="Signupform-content">
-      {getAadhaarForm(state)}
-    </div> */}
-  
-    <div className="Signupform-button-wrapper">
-      <button
-        onClick={() => {
-        state.scanAadhaarbuttonClicked();
-        }}
-        className="Signupform-button"
-      >
-        <span className="Signupform-button-label">
-      Scan Aadhaar Card
-        </span>
-      </button>
-    </div>
-    <div className="Error-message">{state.state.aadhaarError}</div>
-  </div>
-);
-const verifyOTP = state => (
-  <div className="Signupform-welcome-message">
-    <div className="Signupform-heading">
-    An otp has been sent on your registered mobile number
-    </div>
-    <div className="Signupform-content">
-      {verifyOTPForm(state)}
-    </div>
-    <div className="Signupform-button-wrapper">
-      <button
-        className="Signupform-button"
-        onClick={() => state.sendOTPButtonClicked()}
-      >
-        <span className="Signupform-button-label">
-      Resend otp
-        </span>
-      </button>
-      <button
-        onClick={() => {
-        state.verifyOTPButtonClicked();
-      }}
-        className="Signupform-button"
-      >
-        <span className="Signupform-button-label">
-      Verify otp
-        </span>
-      </button>
-    </div>
-  </div>
-);
-
-const getDetails = state => (
-  <div className="Signupform-welcome-message">
-    <div className="Signupform-heading">
-    Verify your details
-    </div>
-    <div className="Signupform-content">
-      <div className="Signupform-verify-details">
-        <div className="Signupform-each-entry">
-          <div className="Signupform-table-label">
-            <span className="Signupform-table-text">
-            Name
-            </span>
-          </div>
-          <div className="Signupform-table-value">
-            <span className="Signupform-table-text">
-              {state.state.eKYCResponse.e_Kyc.Poi.Name}
-            </span>
-          </div>
-        </div>
-        <div className="Signupform-each-entry">
-          <div className="Signupform-table-label">
-            <span className="Signupform-table-text">
-            Gender
-            </span>
-          </div>
-          <div className="Signupform-table-value">
-            <span className="Signupform-table-text">
-              {state.state.eKYCResponse.e_Kyc.Poi.Gender}
-            </span>
-          </div>
-        </div>
-        <div className="Signupform-each-entry">
-          <div className="Signupform-table-label">
-            <span className="Signupform-table-text">
-            DOB
-            </span>
-          </div>
-          <div className="Signupform-table-value">
-            <span className="Signupform-table-text">
-              {strftime('%F', new Date(state.state.eKYCResponse.e_Kyc.Poi.Dob))}
-            </span>
-          </div>
-        </div>
-        <div className="Signupform-each-entry">
-          <div className="Signupform-table-label">
-            <span className="Signupform-table-text">
-            Contact
-            </span>
-          </div>
-          <div className="Signupform-table-value">
-            <span className="Signupform-table-text">
-              {state.state.eKYCResponse.e_Kyc.Poi.contact}
-            </span>
-          </div>
-        </div>
-        <div className="Signupform-each-entry">
-          <div className="Signupform-table-label">
-            <span className="Signupform-table-text">
-            Guardian
-            </span>
-          </div>
-          <div className="Signupform-table-value">
-            <span className="Signupform-table-text">
-              {state.state.eKYCResponse.e_Kyc.Poa.co}
-            </span>
-          </div>
-        </div>
-        <div className="Signupform-each-entry">
-          <div className="Signupform-table-label">
-            <span className="Signupform-table-text">
-            House no.
-            </span>
-          </div>
-          <div className="Signupform-table-value">
-            <span className="Signupform-table-text">
-              {state.state.eKYCResponse.e_Kyc.Poa.house}
-            </span>
-          </div>
-        </div>
-        <div className="Signupform-each-entry">
-          <div className="Signupform-table-label">
-            <span className="Signupform-table-text">
-            Street
-            </span>
-          </div>
-          <div className="Signupform-table-value">
-            <span className="Signupform-table-text">
-              {state.state.eKYCResponse.e_Kyc.Poa.street}
-            </span>
-          </div>
-        </div>
-        <div className="Signupform-each-entry">
-          <div className="Signupform-table-label">
-            <span className="Signupform-table-text">
-            Locality
-            </span>
-          </div>
-          <div className="Signupform-table-value">
-            <span className="Signupform-table-text">
-              {state.state.eKYCResponse.e_Kyc.Poa.lc}
-            </span>
-          </div>
-        </div>
-        <div className="Signupform-each-entry">
-          <div className="Signupform-table-label">
-            <span className="Signupform-table-text">
-            Landmark
-            </span>
-          </div>
-          <div className="Signupform-table-value">
-            <span className="Signupform-table-text">
-              {state.state.eKYCResponse.e_Kyc.Poa.landmark}
-            </span>
-          </div>
-        </div>
-        <div className="Signupform-each-entry">
-          <div className="Signupform-table-label">
-            <span className="Signupform-table-text">
-            Sub district
-            </span>
-          </div>
-          <div className="Signupform-table-value">
-            <span className="Signupform-table-text">
-              {state.state.eKYCResponse.e_Kyc.Poa.subdist}
-            </span>
-          </div>
-        </div>
-        <div className="Signupform-each-entry">
-          <div className="Signupform-table-label">
-            <span className="Signupform-table-text">
-            District
-            </span>
-          </div>
-          <div className="Signupform-table-value">
-            <span className="Signupform-table-text">
-              {state.state.eKYCResponse.e_Kyc.Poa.dist}
-            </span>
-          </div>
-        </div>
-        <div className="Signupform-each-entry">
-          <div className="Signupform-table-label">
-            <span className="Signupform-table-text">
-            Pincode
-            </span>
-          </div>
-          <div className="Signupform-table-value">
-            <span className="Signupform-table-text">
-              {state.state.eKYCResponse.e_Kyc.Poa.pc}
-            </span>
-          </div>
-        </div>
-        <div className="Signupform-each-entry">
-          <div className="Signupform-table-label">
-            <span className="Signupform-table-text">
-            State
-            </span>
-          </div>
-          <div className="Signupform-table-value">
-            <span className="Signupform-table-text">
-              {state.state.eKYCResponse.e_Kyc.Poa.state}
-            </span>
-          </div>
-        </div>
-      </div>
-    </div>
-    <div className="Signupform-button-wrapper">
-      <button
-        className="Signupform-button"
-        onClick={() => {
-          state.setState({
-            notmydetailsOTPError: 'Please get your details verified from nearest aadhaar centre',
-          });
-      }}
-      >
-        <span className="Signupform-button-label">
-   This is not me
-        </span>
-      </button>
-
-      <button
-        onClick={() => {
-        state.detailsVerifiedButtonClicked();
-      }}
-        className="Signupform-button"
-      >
-        <span className="Signupform-button-label">
-      This is me
-        </span>
-      </button>
-    </div>
-    <div className="Error-message">{state.state.notmydetailsOTPError}</div>
-  </div>
-);
-
-const userRegister = state => (
-  <div className="Signupform-welcome-message">
-    <div className="Signupform-heading">Set username and password</div>
-    <div className="Signupform-content">
-      {userRegisterForm(state)}
-    </div>
-    <div className="Signupform-button-wrapper">
-      <button className="Signupform-button" onClick={() => { state.onRegister(); }}>
-        <span className="Signupform-button-label">
-      Open account
-        </span>
-      </button>
-    </div>
-  </div>
-);
 
 class SignupForm extends React.Component {
   constructor() {
@@ -385,7 +36,6 @@ class SignupForm extends React.Component {
   }
   onRegister() {
     axios.post('/users', {
-    //  aadhaarNo:'123412341234',
       aadhaarNo: this.state.aadhaarNo,
       isVerified: this.state.isVerified,
       userName: this.state.username,
@@ -425,7 +75,6 @@ class SignupForm extends React.Component {
   }
 
  scanAadhaarbuttonClicked(){ 
-  // console.log(this.state.aadhaarNo);
     this.setComponent(5);
   }
   useStateError() {
@@ -452,7 +101,6 @@ class SignupForm extends React.Component {
         }, () => {
           console.log('Hello', this.state.eKYCResponse);
         });
-        console.log('aGAIN HELLO', this.state.eKYCResponse);
       } else if (response.data.message === 'Authentication failed') {
         this.setState({
           otpClass: 'error',
@@ -525,8 +173,7 @@ class SignupForm extends React.Component {
       console.log(err);
       if (err.response.status === 400) {
         this.setState({
-          aadhaarError: 'Invalid aadhaar number',
-          
+          aadhaarError: 'Invalid aadhaar number',   
           noOfComponent:1,
         });
       }
@@ -557,8 +204,8 @@ class SignupForm extends React.Component {
         <div className="Signupform-container">
           <div className="Signupform-box">
             <div className="Signupform-component">
-              {progressBarHalf()}
-              {sendOTP(this)}
+              {helpers.progressBarHalf()}
+              {helpers.sendOTP(this)}
             </div>
           </div>
         </div>
@@ -568,12 +215,12 @@ class SignupForm extends React.Component {
         <div className="Signupform-container">
           <div className="Signupform-box">
             <div className="Signupform-component">
-              {progressBar()}
-              {sendOTP(this)}
+              {helpers.progressBar()}
+              {helpers.sendOTP(this)}
             </div>
             <div className="Signupform-component">
-              {progressBarHalf()}
-              {verifyOTP(this)}
+              {helpers.progressBarHalf()}
+              {helpers.verifyOTP(this)}
             </div>
           </div>
         </div>
@@ -583,16 +230,16 @@ class SignupForm extends React.Component {
         <div className="Signupform-container">
           <div className="Signupform-box">
             <div className="Signupform-component">
-              {progressBar()}
-              {sendOTP(this)}
+              {helpers.progressBar()}
+              {helpers.sendOTP(this)}
             </div>
             <div className="Signupform-component">
-              {progressBar()}
-              {verifyOTP(this)}
+              {helpers.progressBar()}
+              {helpers.verifyOTP(this)}
             </div>
             <div className="Signupform-component">
-              {progressBarHalf()}
-              {getDetails(this)}
+              {helpers.progressBarHalf()}
+              {helpers.getDetails(this)}
             </div>
           </div>
         </div>
@@ -607,20 +254,20 @@ class SignupForm extends React.Component {
       <div className="Signupform-container">
         <div className="Signupform-box">
           <div className="Signupform-component">
-            {progressBar()}
-            {sendOTP(this)}
+            {helpers.progressBar()}
+            {helpers.sendOTP(this)}
           </div>
           <div className="Signupform-component">
-            {progressBar()}
-            {verifyOTP(this)}
+            {helpers.progressBar()}
+            {helpers.verifyOTP(this)}
           </div>
           <div className="Signupform-component">
-            {progressBar()}
-            {getDetails(this)}
+            {helpers.progressBar()}
+            {helpers.getDetails(this)}
           </div>
           <div className="Signupform-component">
-            {progressBarHalf()}
-            {userRegister(this)}
+            {helpers.progressBarHalf()}
+            {helpers.userRegister(this)}
           </div>
         </div>
       </div>
